livecheck: add user agent option
This adds a `user_agent` option that can be set in `livecheck` blocks
(as a `url` option) to specify the curl user agent. This value is
passed to `Utils::Curl::curl_args`, so the expected value is a string
or supported symbol (`:browser`/`:fake` or `:default`).

In this implementation, an explicit `user_agent` value will be used
in place of the existing `:default` and `:browser` user agents that
are used by default. This retains the existing behavior as the default
for now but I intend to remove the implicit user agent retry logic
once existing usage is replaced by explicit `user_agent` options in
`livecheck` blocks. [The user agent retry logic was originally added
because it wasn't possible to set a user agent in a `livecheck` block
yet and I only accepted it as a temporary measure that would
eventually be removed.]